+++ config/whatwedo-wordpress.php
@@ -3,8 +3,9 @@
 declare(strict_types=1);
 
 use Symplify\EasyCodingStandard\Config\ECSConfig;
+use whatwedo\PhpCodingStandard\Set\WhatwedoSets;
 
 return ECSConfig::configure()
     ->withSets([
-        __DIR__ . '/whatwedo-common.php',
+        WhatwedoSets::COMMON,
     ]);