+++ ecs.php
@@ -3,11 +3,13 @@
 declare(strict_types=1);
 
 use Symplify\EasyCodingStandard\Config\ECSConfig;
+use whatwedo\PhpCodingStandard\Set\WhatwedoSets;
 
 return ECSConfig::configure()
     ->withPaths([
-        __DIR__ . '/',
+        __DIR__ . '/config',
+        __DIR__ . '/src',
     ])
     ->withSets([
-        __DIR__ . '/config/whatwedo-common.php',
+        WhatwedoSets::COMMON,
     ]);